Docs: add ability to expand code examples (#1230)
While scrolling through our component docs, the amount of code we see is overwhelming. This PR adds an `Expand code` button which expands the code when required.

![gestalt-docs-expand-code-exampe](https://user-images.githubusercontent.com/127199/93911571-6ca21e00-fcb7-11ea-9273-54656f2b4814.gif)

### Notes
* The max height of the code is `100` with a threshold of `80`. So we will only show the expand button when the height is > 180px.
* This is a first stab - definitely open for design suggestions.

## Test Plan

* On the netlify preview, we see `Expand code` on all code examples which have a height of + 180px.
